컨트롤러 생성 install 라우터에 대한 컨트롤러 생성
